requirements.txt
pandas
plotly_express
streamlit
numpy
scikit-learn